feat/#85: 일별 혈당 목록 조회 api 개발, comparator 개발

+++ src/main/java/com/epi/epilog/app/dto/LogsResponseDto.java
@@ -90,7 +90,7 @@ public class LogsResponseDto {
     @NoArgsConstructor(access =  AccessLevel.PRIVATE)
     public static class DayBloodSugarItem {
         private String title;
-        private Integer bloodSugar;
+        private Double bloodSugar;
     }
 
     /**
